+++ RacingCar/main.swift
@@ -6,6 +6,5 @@
 
 import Foundation
 
-
 let game = GameController()
 game.gameStart()
